chat-bot/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-chroma
langchain-google-genai
langchain-community # Add this line
beautifulsoup4